pulsar.cx: correct workDir to private
workDir was being set to `/var/cache/github-runner-pulsar-{0..5}` instead of
`/var/cache/private/github-runner-pulsar-{0..5}`. This is a symlink so it was
mostly working, but the GitHub Actions Cache step explicitly creates a path
relative to GITHUB_WORKSPACE and this was messing up the paths. This meant
caches failed to create on this machine, though they could load fine. Add in the
`/private` explicitly so these relative paths are accessible.

Test plan:
- Ran on https://github.com/sched-ext/scx/pull/1705 which failed before and
  passes now.