pulsar.cx: add signing key and trust it
We don't have root on this machine in the GitHub Actions runners so can't load
untrusted caches into the store. We could risk not using the cache at all and
relying on the entry already being on this machine, but it would be better if
we can load them.

Sign all cache results to allow loading untrusted caches into the store.

Test plan:
- Will run the failing workflow with this and see if it helps.