pulsar.cx: add cachix cache
We don't have root on this machine in the GitHub Actions runners so can't load
untrusted caches into the store. We could risk not using the cache at all and
relying on the entry already being on this machine, but it would be better if
we can load them.

Add the cachix cache as a substituter and trust its key.

Test plan:
- Ran the cachix workflow. It works well.